requirements-dev.txt
langgraph
langchain
langchain-ollama
langchain-openai
pydantic
PyYAML
SQLAlchemy
psycopg[binary]
python-dotenv
alembic
